refactor: flatten directory structure from core/api/core/api to core/api
