--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/api/config/AsyncConfig.kt
@@ -1,27 +1,0 @@
-package noweekend.core.api.core.api.config
-
-import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler
-import org.springframework.context.annotation.Configuration
-import org.springframework.scheduling.annotation.AsyncConfigurer
-import org.springframework.scheduling.annotation.EnableAsync
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor
-import java.util.concurrent.Executor
-
-@Configuration
-@EnableAsync
-class AsyncConfig : AsyncConfigurer {
-    override fun getAsyncExecutor(): Executor {
-        val executor = ThreadPoolTaskExecutor()
-        executor.corePoolSize = 10
-        executor.maxPoolSize = 10
-        executor.queueCapacity = 10000
-        executor.setWaitForTasksToCompleteOnShutdown(true)
-        executor.setAwaitTerminationSeconds(10)
-        executor.initialize()
-        return executor
-    }
-
-    override fun getAsyncUncaughtExceptionHandler(): AsyncUncaughtExceptionHandler {
-        return AsyncExceptionHandler()
-    }
-}
