--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/api/config/AsyncExceptionHandler.kt
@@ -1,23 +1,0 @@
-package noweekend.core.api.core.api.config
-
-import noweekend.core.api.core.support.error.CoreException
-import org.slf4j.LoggerFactory
-import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler
-import org.springframework.boot.logging.LogLevel
-import java.lang.reflect.Method
-
-class AsyncExceptionHandler : AsyncUncaughtExceptionHandler {
-    private val log = LoggerFactory.getLogger(javaClass)
-
-    override fun handleUncaughtException(e: Throwable, method: Method, vararg params: Any?) {
-        if (e is CoreException) {
-            when (e.errorType.logLevel) {
-                LogLevel.ERROR -> log.error("CoreException : {}", e.message, e)
-                LogLevel.WARN -> log.warn("CoreException : {}", e.message, e)
-                else -> log.info("CoreException : {}", e.message, e)
-            }
-        } else {
-            log.error("Exception : {}", e.message, e)
-        }
-    }
-}
