--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/api/controller/ApiControllerAdvice.kt
@@ -1,32 +1,0 @@
-package noweekend.core.api.core.api.controller
-
-import noweekend.core.api.core.support.error.CoreException
-import noweekend.core.api.core.support.error.ErrorType
-import noweekend.core.api.core.support.response.ApiResponse
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
-import org.springframework.boot.logging.LogLevel
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.ExceptionHandler
-import org.springframework.web.bind.annotation.RestControllerAdvice
-
-@RestControllerAdvice
-class ApiControllerAdvice {
-    private val log: Logger = LoggerFactory.getLogger(javaClass)
-
-    @ExceptionHandler(CoreException::class)
-    fun handleCoreException(e: CoreException): ResponseEntity<ApiResponse<Any>> {
-        when (e.errorType.logLevel) {
-            LogLevel.ERROR -> log.error("CoreException : {}", e.message, e)
-            LogLevel.WARN -> log.warn("CoreException : {}", e.message, e)
-            else -> log.info("CoreException : {}", e.message, e)
-        }
-        return ResponseEntity(ApiResponse.error(e.errorType, e.data), e.errorType.status)
-    }
-
-    @ExceptionHandler(Exception::class)
-    fun handleException(e: Exception): ResponseEntity<ApiResponse<Any>> {
-        log.error("Exception : {}", e.message, e)
-        return ResponseEntity(ApiResponse.error(ErrorType.DEFAULT_ERROR), ErrorType.DEFAULT_ERROR.status)
-    }
-}
