--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/api/controller/v1/response/ExampleResponseDto.kt
@@ -1,5 +1,0 @@
-package noweekend.core.api.core.api.controller.v1.response
-
-data class ExampleResponseDto(
-    val result: String,
-)
