--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/domain/ExampleData.kt
@@ -1,6 +1,0 @@
-package noweekend.core.api.core.domain
-
-data class ExampleData(
-    val value: String,
-    val param: String,
-)
