--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/domain/ExampleResult.kt
@@ -1,5 +1,0 @@
-package noweekend.core.api.core.domain
-
-data class ExampleResult(
-    val data: String,
-)
