--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/domain/ExampleService.kt
@@ -1,10 +1,0 @@
-package noweekend.core.api.core.domain
-
-import org.springframework.stereotype.Service
-
-@Service
-class ExampleService() {
-    fun processExample(exampleData: ExampleData): ExampleResult {
-        return ExampleResult(exampleData.value)
-    }
-}
