--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/support/error/CoreException.kt
@@ -1,6 +1,0 @@
-package noweekend.core.api.core.support.error
-
-class CoreException(
-    val errorType: ErrorType,
-    val data: Any? = null,
-) : RuntimeException(errorType.message)
