--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/support/error/ErrorCode.kt
@@ -1,6 +1,0 @@
-package noweekend.core.api.core.support.error
-
-enum class ErrorCode {
-    E401,
-    E500,
-}
