--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/support/error/ErrorMessage.kt
@@ -1,13 +1,0 @@
-package noweekend.core.api.core.support.error
-
-data class ErrorMessage private constructor(
-    val code: String,
-    val message: String,
-    val data: Any? = null,
-) {
-    constructor(errorType: ErrorType, data: Any? = null) : this(
-        code = errorType.code.name,
-        message = errorType.message,
-        data = data,
-    )
-}
