--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/support/response/ApiResponse.kt
@@ -1,24 +1,0 @@
-package noweekend.core.api.core.support.response
-
-import noweekend.core.api.core.support.error.ErrorMessage
-import noweekend.core.api.core.support.error.ErrorType
-
-data class ApiResponse<T> private constructor(
-    val result: ResultType,
-    val data: T? = null,
-    val error: ErrorMessage? = null,
-) {
-    companion object {
-        fun success(): ApiResponse<Any> {
-            return ApiResponse(ResultType.SUCCESS, null, null)
-        }
-
-        fun <S> success(data: S): ApiResponse<S> {
-            return ApiResponse(ResultType.SUCCESS, data, null)
-        }
-
-        fun <S> error(error: ErrorType, errorData: Any? = null): ApiResponse<S> {
-            return ApiResponse(ResultType.ERROR, null, ErrorMessage(error, errorData))
-        }
-    }
-}
