--- noweekend-core/core-api/src/main/kotlin/noweekend/core/api/core/support/response/ResultType.kt
@@ -1,6 +1,0 @@
-package noweekend.core.api.core.support.response
-
-enum class ResultType {
-    SUCCESS,
-    ERROR,
-}
