+++ noweekend-core/core-api/src/main/kotlin/noweekend/core/api/security/jwt/JwtProviderImpl.kt
@@ -1,9 +1,9 @@
-package noweekend.core.api.core.api.security.jwt
+package noweekend.core.api.security.jwt
 
 import io.jsonwebtoken.Jwts
 import io.jsonwebtoken.SignatureAlgorithm
 import io.jsonwebtoken.security.Keys
-import noweekend.core.api.core.api.config.JwtProperties
+import noweekend.core.api.config.JwtProperties
 import org.springframework.stereotype.Component
 import java.util.Date
