+++ noweekend-core/core-api/src/main/kotlin/noweekend/core/support/error/ErrorCode.kt
@@ -1,5 +1,6 @@
 package noweekend.core.support.error
 
 enum class ErrorCode {
+    E401,
     E500,
 }
