+++ noweekend-core/core-api/src/main/kotlin/noweekend/core/support/error/ErrorType.kt
@@ -4,5 +4,6 @@ import org.springframework.boot.logging.LogLevel
 import org.springframework.http.HttpStatus
 
 enum class ErrorType(val status: HttpStatus, val code: ErrorCode, val message: String, val logLevel: LogLevel) {
+    UNAUTHORIZED(HttpStatus.UNAUTHORIZED, ErrorCode.E401, "Unauthorized user", LogLevel.DEBUG),
     DEFAULT_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, ErrorCode.E500, "An unexpected error has occurred.", LogLevel.ERROR),
 }
